Wrote a test that will read data on mcu1 and send it to mcu2 for verification.
Still trying to narrow down the issue with the dataline reflections. Tried terminating resistors on the data legs and clk pin but it didn't seem to fix any of the issues to mcu1. MCU2 WAS able to read at 133mhz with some 330ohm resistors on u4 and u8 and a drive strength of 8mA