IMPORTANT:: pick-sdk crt0.S was modified to speed up rp2040 boot time. _reset_handler:     ldr r0, =(ROSC_BASE + ROSC_DIV_OFFSET)     ldr r1, =0xaa1     str r1, [r0]
Overclock rp2040s to 360MHz (they get warm but seem to be okay)
Allows full speed rom booting.

Moved some files in the menu rom so they don't compile into the rom and bloat the size.

Added in gameCic to menu rom so when it loads a new game it can properly set the cic of the rom being loaded.
TODO this needs to be setup so it will properly fetch the gameCIC, it's hard coded atm to 5(6105).

Fixed but with selected rom title length. Not sure what happened but using strncpy now as well as the actual length of the string sent from the n64.