Added a delay before loading box art in the menu rom. Helps it feel better when scrolling.
This will compile and boot to menu rom. Bus timing is set to 0x2040, clocks are 336/4, although it likely doesn't need to run that fast.

Eeprom save size is being fetched and sent to the mcu1. There seems to be an issue loading eeprom saves, though originally I thought it was because the file was in a sub folder. Not sure if there is a memory issue or what. Might only be able to have one file open at a time? Unsure but it requires more testing.